implement force delete for inventory

+++ src/main/java/seedu/address/logic/commands/inventory/InventoryDeleteCommand.java
@@ -10,7 +10,9 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
+import seedu.address.model.dish.Dish;
 import seedu.address.model.ingredient.Ingredient;
+import seedu.address.model.order.Order;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
@@ -26,11 +28,28 @@ public class InventoryDeleteCommand extends Command {
             + "Example: " + COMPONENT_WORD + " " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_DELETE_INGREDIENT_SUCCESS = "Deleted ingredient: %1$s";
+    public static final String MESSAGE_DELETE_INGREDIENT_FAILURE =
+            "Failed to deleted ingredient: %1$s due to associated dishes, "
+                    + "add -f flag to force delete the ingredient\n"
+                    + "Warning: This will delete any dishes that contains %1$s\n"
+                    + "\t\t   This will also mark associated orders as 'Cancelled'";
 
     private final Index targetIndex;
+    private final boolean isForce;
 
     public InventoryDeleteCommand(Index targetIndex) {
         this.targetIndex = targetIndex;
+        this.isForce = false;
+    }
+
+    /**
+     * Forces delete dish with given index along with associated ingredients
+     * @param targetIndex index in ingredient list
+     * @param isForce forces delete a dish
+     */
+    public InventoryDeleteCommand(Index targetIndex, boolean isForce) {
+        this.targetIndex = targetIndex;
+        this.isForce = isForce;
     }
 
     @Override
@@ -45,6 +64,19 @@ public class InventoryDeleteCommand extends Command {
         }
 
         Ingredient ingredientToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        List<Dish> dishesToCascade = model.getDishesByIngredients(ingredientToDelete);
+        boolean dishesContainsIngredient = !dishesToCascade.isEmpty();
+
+        if (dishesContainsIngredient && !isForce) {
+            throw new CommandException(String.format(MESSAGE_DELETE_INGREDIENT_FAILURE, ingredientToDelete.getName()));
+        }
+
+        for(Dish dishToDelete: dishesToCascade){
+            List<Order> outstandingOrders = model.getIncompleteOrdersContainingDish(dishToDelete);
+            model.cancelOrders(outstandingOrders);
+            model.deleteDish(dishToDelete);
+        }
         model.deleteIngredient(ingredientToDelete);
 
         // Delete dish here from model
